Install rocm scripts
