+++ tools/cuda_utils.py
@@ -1,14 +1,11 @@
-import importlib
+import argparse
 import os
 import re
 import subprocess
 from pathlib import Path
 
-from typing import Optional
-
 # defines the default CUDA version to compile against
 DEFAULT_CUDA_VERSION = "12.4"
-REPO_ROOT = Path(__file__).parent.parent
 
 CUDA_VERSION_MAP = {
     "12.4": {
@@ -17,10 +14,6 @@ CUDA_VERSION_MAP = {
         "jax": "jax[cuda12]",
     },
 }
-PIN_CMAKE_VERSION = "3.22.*"
-
-TORCH_NIGHTLY_PACKAGES = ["torch"]
-BUILD_REQUIREMENTS_FILE = REPO_ROOT.joinpath("utils", "build_requirements.txt")
 
 
 def _nvcc_output_match(nvcc_output, target_cuda_version):
@@ -94,6 +87,7 @@ def setup_cuda_softlink(cuda_version: str):
 
 
 def install_pytorch_nightly(cuda_version: str, env, dryrun=False):
+    from .torch.utils import TORCH_NIGHTLY_PACKAGES
     uninstall_torch_cmd = ["pip", "uninstall", "-y"]
     uninstall_torch_cmd.extend(TORCH_NIGHTLY_PACKAGES)
     if dryrun:
@@ -131,68 +125,7 @@ def install_torch_deps(cuda_version: str):
     subprocess.check_call(cmd)
 
 
-def install_torch_build_deps(cuda_version: str):
-    install_torch_deps(cuda_version=cuda_version)
-    # Pin cmake version to stable
-    # See: https://github.com/pytorch/builder/pull/1269
-    torch_build_deps = [
-        "cffi",
-        "sympy",
-        "typing_extensions",
-        "future",
-        "six",
-        "dataclasses",
-        "tabulate",
-        "tqdm",
-        "mkl",
-        "mkl-include",
-        f"cmake={PIN_CMAKE_VERSION}",
-    ]
-    cmd = ["conda", "install", "-y"] + torch_build_deps
-    subprocess.check_call(cmd)
-    build_deps = ["ffmpeg"]
-    cmd = ["conda", "install", "-y"] + build_deps
-    subprocess.check_call(cmd)
-    # pip build deps
-    cmd = ["pip", "install", "-r"] + [str(BUILD_REQUIREMENTS_FILE.resolve())]
-    subprocess.check_call(cmd)
-    # conda forge deps
-    # ubuntu 22.04 comes with libstdcxx6 12.3.0
-    # we need to install the same library version in conda to maintain ABI compatibility
-    conda_deps = ["libstdcxx-ng=12.3.0"]
-    cmd = ["conda", "install", "-y", "-c", "conda-forge"] + conda_deps
-    subprocess.check_call(cmd)
-
-
-def get_torch_nightly_version(pkg_name: str):
-    pkg = importlib.import_module(pkg_name)
-    version = pkg.__version__
-    regex = ".*dev([0-9]+).*"
-    date_str = re.match(regex, version).groups()[0]
-    pkg_ver = {"version": version, "date": date_str}
-    return (pkg_name, pkg_ver)
-
-
-def check_torch_nightly_version(force_date: Optional[str] = None):
-    pkg_versions = dict(map(get_torch_nightly_version, TORCH_NIGHTLY_PACKAGES))
-    pkg_dates = [x[1]["date"] for x in pkg_versions.items()]
-    if not len(set(pkg_dates)) == 1:
-        raise RuntimeError(
-            f"Found more than 1 dates in the torch nightly packages: {pkg_versions}."
-        )
-    if force_date and not pkg_dates[0] == force_date:
-        raise RuntimeError(
-            f"Force date value {force_date}, but found torch packages {pkg_versions}."
-        )
-    force_date_str = f"User force date {force_date}" if force_date else ""
-    print(
-        f"Installed consistent torch nightly packages: {pkg_versions}. {force_date_str}"
-    )
-
-
 if __name__ == "__main__":
-    import argparse
-
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "--cudaver",
@@ -234,9 +167,12 @@ if __name__ == "__main__":
     if args.install_torch_deps:
         install_torch_deps(cuda_version=args.cudaver)
     if args.install_torch_build_deps:
-        install_torch_build_deps(cuda_version=args.cudaver)
+        from .torch.utils import install_torch_build_deps
+        install_torch_deps(cuda_version=args.cudaver)
+        install_torch_build_deps()
     if args.install_torch_nightly:
         install_pytorch_nightly(cuda_version=args.cudaver, env=os.environ)
     if args.check_torch_nightly_version:
+        from .torch.utils import check_torch_nightly_version
         assert not args.install_torch_nightly, "Error: Can't run install torch nightly and check version in the same command."
         check_torch_nightly_version(args.force_date)
