+++ tools/rocm_utils.py
@@ -1,0 +1,3 @@
+
+# defines the default ROCM version to compile against
+DEFAULT_ROCM_VERSION = "6.2"
